Fix process hash mismatch between 32-bit and 64-bit nodes.
Hash iterable sizes as uint64_t instead of native size_t so spawn/trace indices agree across architectures.

+++ src/lib/common/serialize.hpp
@@ -395,9 +395,10 @@ namespace details {
         return s;
     }
 
+    //! @brief Hash iterable size as uint64_t so 32-bit and 64-bit nodes agree.
     template <typename T, typename S>
     hstream& iterable_serialize(hstream& s, T& x, S, wrapper<void> = {}) {
-        s.write(x.size());
+        s.write(static_cast<uint64_t>(x.size()));
         for (auto& i : x) s & i;
         return s;
     }
